+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-0.19 (2024-06-??)
+0.19 (2024-06-09)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/9)
 * Optimialization using memo, useMemo and useCallback